requirements.txt
pydantic~=2.10.6
psycopg2~=2.9.10
python-dotenv~=1.0.1
fastapi~=0.115.7
uvicorn~=0.34.0
email-validator~=2.2.0
logger~=1.4
dnspython~=2.7.0
idna~=3.10
anyio~=4.8.0
h11~=0.14.0
pip~=24.3.1
sniffio~=1.3.1
typing_extensions~=4.12.2
click~=8.1.8
starlette~=0.45.3
email_validator~=2.2.0
pydantic_core~=2.27.2
annotated-types~=0.7.0
sqlmodel~=0.0.22